djangoapp/requirements.txt
asgiref>=3.11.1, <3.12
Django>=6.0.2, <6.1
django-cors-headers>=4.9.0, <4.10
djangorestframework>=3.16.1, <3.17
djangorestframework_simplejwt>=5.5.1, <5.6
pillow>=12.1.1, <12.2
PyJWT>=2.11.0, <2.12
sqlparse>=0.5.5, <0.6
tzdata>=2025.3, <2025.4
